els-ASAP auto commit file/tag nvme-cli-1.3-1.el7.src.rpm

+++ src/suffix.h
@@ -31,11 +31,9 @@
 
 #ifndef __ARGCONFIG_SUFFIX_H__
 
-#include <inttypes.h>
-
 const char *suffix_si_get(double *value);
 const char *suffix_binary_get(long long *value);
 const char *suffix_dbinary_get(double *value);
-uint64_t suffix_binary_parse(const char *value);
+long long suffix_binary_parse(const char *value);
 
 #endif
